refactor: move slice related functions
